Add a flag indicating the first run in docker.

+++ entrypoint.sh
@@ -1,32 +1,35 @@
 #!/usr/bin/env bash
 
-# 如不分离备份的 github 账户，默认与哪吒登陆的 github 账户一致
-GH_BACKUP_USER=${GH_BACKUP_USER:-$GH_USER}
-WORK_DIR=/dashboard
+# 首次运行时执行以下流程，再次运行时存在 /etc/supervisor/conf.d/damon.conf 文件，直接到最后一步
+if [ ! -s /etc/supervisor/conf.d/damon.conf ]; then
 
-error() { echo -e "\033[31m\033[01m$*\033[0m" && exit 1; } # 红色
-info() { echo -e "\033[32m\033[01m$*\033[0m"; }   # 绿色
-hint() { echo -e "\033[33m\033[01m$*\033[0m"; }   # 黄色
+  # 如不分离备份的 github 账户，默认与哪吒登陆的 github 账户一致
+  GH_BACKUP_USER=${GH_BACKUP_USER:-$GH_USER}
+  WORK_DIR=/dashboard
 
-# 如参数不齐全，容器退出，另外处理某些环境变量填错后的处理
-[[ -z "$GH_USER" || -z "$GH_CLIENTID" || -z "$GH_CLIENTSECRET" || -z "$ARGO_AUTH" || -z "$ARGO_DOMAIN" ]] && error " There are variables that are not set. "
-[[ "$ARGO_AUTH" =~ TunnelSecret ]] && grep -qv '"' <<< "$ARGO_AUTH" && ARGO_AUTH=$(sed 's@{@{"@g;s@[,:]@"\0"@g;s@}@"}@g' <<< "$ARGO_AUTH")  # Json 时，没有了"的处理
-[[ "$ARGO_AUTH" =~ ey[A-Z0-9a-z=]{120,250}$ ]] && ARGO_AUTH=$(awk '{print $NF}' <<< "$ARGO_AUTH") # Token 复制全部，只取最后的 ey 开始的
-[ -n "$GH_REPO" ] && grep -q '/' <<< "$GH_REPO" && GH_REPO=$(awk -F '/' '{print $NF}' <<< "$GH_REPO")  # 填了项目全路径的处理
+  error() { echo -e "\033[31m\033[01m$*\033[0m" && exit 1; } # 红色
+  info() { echo -e "\033[32m\033[01m$*\033[0m"; }   # 绿色
+  hint() { echo -e "\033[33m\033[01m$*\033[0m"; }   # 黄色
 
-echo -e "nameserver 127.0.0.11\nnameserver 8.8.4.4\nnameserver 223.5.5.5\nnameserver 2001:4860:4860::8844\nnameserver 2400:3200::1\n" > /etc/resolv.conf
+  # 如参数不齐全，容器退出，另外处理某些环境变量填错后的处理
+  [[ -z "$GH_USER" || -z "$GH_CLIENTID" || -z "$GH_CLIENTSECRET" || -z "$ARGO_AUTH" || -z "$ARGO_DOMAIN" ]] && error " There are variables that are not set. "
+  [[ "$ARGO_AUTH" =~ TunnelSecret ]] && grep -qv '"' <<< "$ARGO_AUTH" && ARGO_AUTH=$(sed 's@{@{"@g;s@[,:]@"\0"@g;s@}@"}@g' <<< "$ARGO_AUTH")  # Json 时，没有了"的处理
+  [[ "$ARGO_AUTH" =~ ey[A-Z0-9a-z=]{120,250}$ ]] && ARGO_AUTH=$(awk '{print $NF}' <<< "$ARGO_AUTH") # Token 复制全部，只取最后的 ey 开始的
+  [ -n "$GH_REPO" ] && grep -q '/' <<< "$GH_REPO" && GH_REPO=$(awk -F '/' '{print $NF}' <<< "$GH_REPO")  # 填了项目全路径的处理
 
-# 下载需要的应用
-wget -c https://github.com/fscarmen2/Argo-Nezha-Service-Container/releases/download/grpcwebproxy/grpcwebproxy_linux_$(uname -m | sed "s#x86_64#amd64#; s#aarch64#arm64#").tar.gz -qO- | tar xz -C $WORK_DIR
-wget -qO $WORK_DIR/cloudflared https://github.com/cloudflare/cloudflared/releases/latest/download/cloudflared-linux-$(uname -m | sed "s#x86_64#amd64#; s#aarch64#arm64#")
-wget -O $WORK_DIR/nezha-agent.zip https://github.com/nezhahq/agent/releases/latest/download/nezha-agent_linux_$(uname -m | sed "s#x86_64#amd64#; s#aarch64#arm64#").zip
-unzip $WORK_DIR/nezha-agent.zip -d $WORK_DIR/
+  echo -e "nameserver 127.0.0.11\nnameserver 8.8.4.4\nnameserver 223.5.5.5\nnameserver 2001:4860:4860::8844\nnameserver 2400:3200::1\n" > /etc/resolv.conf
 
-rm -f $WORK_DIR/nezha-agent.zip
+  # 下载需要的应用
+  wget -c https://github.com/fscarmen2/Argo-Nezha-Service-Container/releases/download/grpcwebproxy/grpcwebproxy_linux_$(uname -m | sed "s#x86_64#amd64#; s#aarch64#arm64#").tar.gz -qO- | tar xz -C $WORK_DIR
+  wget -qO $WORK_DIR/cloudflared https://github.com/cloudflare/cloudflared/releases/latest/download/cloudflared-linux-$(uname -m | sed "s#x86_64#amd64#; s#aarch64#arm64#")
+  wget -O $WORK_DIR/nezha-agent.zip https://github.com/nezhahq/agent/releases/latest/download/nezha-agent_linux_$(uname -m | sed "s#x86_64#amd64#; s#aarch64#arm64#").zip
+  unzip $WORK_DIR/nezha-agent.zip -d $WORK_DIR/
 
-# 根据参数生成哪吒服务端配置文件
-[ ! -d data ] && mkdir data
-cat > ${WORK_DIR}/data/config.yaml << EOF
+  rm -f $WORK_DIR/nezha-agent.zip
+
+  # 根据参数生成哪吒服务端配置文件
+  [ ! -d data ] && mkdir data
+  cat > ${WORK_DIR}/data/config.yaml << EOF
 debug: false
 httpport: 80
 language: zh-CN
@@ -46,19 +49,19 @@ site:
   theme: "default"
 EOF
 
-# SSH path 与 GH_CLIENTSECRET 一样
-echo root:"$GH_CLIENTSECRET" | chpasswd root
-sed -i 's/^#\?PermitRootLogin.*/PermitRootLogin yes/g;s/^#\?PasswordAuthentication.*/PasswordAuthentication yes/g' /etc/ssh/sshd_config
-service ssh restart
+  # SSH path 与 GH_CLIENTSECRET 一样
+  echo root:"$GH_CLIENTSECRET" | chpasswd root
+  sed -i 's/^#\?PermitRootLogin.*/PermitRootLogin yes/g;s/^#\?PasswordAuthentication.*/PasswordAuthentication yes/g' /etc/ssh/sshd_config
+  service ssh restart
 
-# 判断 ARGO_AUTH 为 json 还是 token
-# 如为 json 将生成 argo.json 和 argo.yml 文件
-if [[ "$ARGO_AUTH" =~ TunnelSecret ]]; then
-  ARGO_RUN="cloudflared tunnel --edge-ip-version auto --config $WORK_DIR/argo.yml run"
+  # 判断 ARGO_AUTH 为 json 还是 token
+  # 如为 json 将生成 argo.json 和 argo.yml 文件
+  if [[ "$ARGO_AUTH" =~ TunnelSecret ]]; then
+    ARGO_RUN="cloudflared tunnel --edge-ip-version auto --config $WORK_DIR/argo.yml run"
 
-  echo "$ARGO_AUTH" > $WORK_DIR/argo.json
+    echo "$ARGO_AUTH" > $WORK_DIR/argo.json
 
-  cat > $WORK_DIR/argo.yml << EOF
+    cat > $WORK_DIR/argo.yml << EOF
 tunnel: $(cut -d '"' -f12 <<< "$ARGO_AUTH")
 credentials-file: $WORK_DIR/argo.json
 protocol: http2
@@ -78,18 +81,18 @@ ingress:
   - service: http_status:404
 EOF
 
-# 如为 token 时
-elif [[ "$ARGO_AUTH" =~ ^ey[A-Z0-9a-z=]{120,250}$ ]]; then
-  ARGO_RUN="cloudflared tunnel --edge-ip-version auto --protocol http2 run --token ${ARGO_AUTH}"
-fi
+  # 如为 token 时
+  elif [[ "$ARGO_AUTH" =~ ^ey[A-Z0-9a-z=]{120,250}$ ]]; then
+    ARGO_RUN="cloudflared tunnel --edge-ip-version auto --protocol http2 run --token ${ARGO_AUTH}"
+  fi
 
-# 生成自签署SSL证书
-openssl genrsa -out $WORK_DIR/nezha.key 2048
-openssl req -new -subj "/CN=$ARGO_DOMAIN" -key $WORK_DIR/nezha.key -out $WORK_DIR/nezha.csr
-openssl x509 -req -days 36500 -in $WORK_DIR/nezha.csr -signkey $WORK_DIR/nezha.key -out $WORK_DIR/nezha.pem
+  # 生成自签署SSL证书
+  openssl genrsa -out $WORK_DIR/nezha.key 2048
+  openssl req -new -subj "/CN=$ARGO_DOMAIN" -key $WORK_DIR/nezha.key -out $WORK_DIR/nezha.csr
+  openssl x509 -req -days 36500 -in $WORK_DIR/nezha.csr -signkey $WORK_DIR/nezha.key -out $WORK_DIR/nezha.pem
 
-# 生成定时备份数据库脚本，定时任务，删除 5 天前的备份
-  cat > $WORK_DIR/backup.sh << EOF
+  # 生成定时备份数据库脚本，定时任务，删除 5 天前的备份
+    cat > $WORK_DIR/backup.sh << EOF
 #!/usr/bin/env bash
 
 GH_PAT=$GH_PAT
@@ -184,9 +187,9 @@ fi
 [ \$(supervisorctl status all | grep -c "RUNNING") = \$(grep -c '\[program:.*\]' /etc/supervisor/conf.d/damon.conf) ] && info "\n Done! \n" || error "\n Fail! \n"
 EOF
 
-if [[ -n "$GH_BACKUP_USER" && -n "$GH_EMAIL" && -n "$GH_REPO" && -n "$GH_PAT" ]]; then
-  # 生成还原数据脚本
-  cat > $WORK_DIR/restore.sh << EOF
+  if [[ -n "$GH_BACKUP_USER" && -n "$GH_EMAIL" && -n "$GH_REPO" && -n "$GH_PAT" ]]; then
+    # 生成还原数据脚本
+    cat > $WORK_DIR/restore.sh << EOF
 #!/usr/bin/env bash
 
 # restore.sh 传参 a 自动还原 README.md 记录的文件，当本地与远程记录文件一样时不还原； 传参 f 不管本地记录文件，强制还原成备份库里 README.md 记录的文件； 传参 dashboard-***.tar.gz 还原成备份库里的该文件；不带参数则要求选择备份库里的文件名
@@ -251,7 +254,7 @@ if [ -e \$TEMP_DIR/backup.tar.gz ]; then
 
   # 容器版的备份旧方案是 /dashboard 文件夹，新方案是备份工作目录 < WORK_DIR > 下的文件，此判断用于根据压缩包里的目录架构判断到哪个目录下解压，以兼容新旧备份方案
   FILE_LIST=\$(tar tzf \$TEMP_DIR/backup.tar.gz)
-  FILE_PATH=\$(sed -n 's#\(.*/\)data/sqlite\.db.*#\1#gp' <<< "\$FILE_LIST")  
+  FILE_PATH=\$(sed -n 's#\(.*/\)data/sqlite\.db.*#\1#gp' <<< "\$FILE_LIST")
 
   # 判断备份文件里是否有用户自定义主题，如有则一并解压到临时文件夹
   CUSTOM_PATH=(\$(sed -n "/-custom/s#\$FILE_PATH\(.*-custom\)/.*#\1#gp" <<< "\$FILE_LIST" | sort -u))
@@ -281,14 +284,14 @@ fi
 [ \$(supervisorctl status all | grep -c "RUNNING") = \$(grep -c '\[program:.*\]' /etc/supervisor/conf.d/damon.conf) ] && info "\n Done! \n" || error "\n Fail! \n"
 EOF
 
-  # 生成定时任务，每天北京时间 4:00:00 备份一次，并重启 cron 服务; 每分钟自动检测在线备份文件里的内容
-  grep -q "$WORK_DIR/backup.sh" /etc/crontab || echo "0 4 * * * root bash $WORK_DIR/backup.sh a" >> /etc/crontab
-  grep -q "$WORK_DIR/restore.sh" /etc/crontab || echo "* * * * * root bash $WORK_DIR/restore.sh a" >> /etc/crontab
-  service cron restart
-fi
+    # 生成定时任务，每天北京时间 4:00:00 备份一次，并重启 cron 服务; 每分钟自动检测在线备份文件里的内容
+    grep -q "$WORK_DIR/backup.sh" /etc/crontab || echo "0 4 * * * root bash $WORK_DIR/backup.sh a" >> /etc/crontab
+    grep -q "$WORK_DIR/restore.sh" /etc/crontab || echo "* * * * * root bash $WORK_DIR/restore.sh a" >> /etc/crontab
+    service cron restart
+  fi
 
-# 生成 supervisor 进程守护配置文件
-cat > /etc/supervisor/conf.d/damon.conf << EOF
+  # 生成 supervisor 进程守护配置文件
+  cat > /etc/supervisor/conf.d/damon.conf << EOF
 [supervisord]
 nodaemon=true
 logfile=/dev/null
@@ -323,8 +326,10 @@ stderr_logfile=/dev/null
 stdout_logfile=/dev/null
 EOF
 
-# 赋执行权给 sh 及所有应用
-chmod +x $WORK_DIR/{grpcwebproxy,cloudflared,nezha-agent,*.sh}
+  # 赋执行权给 sh 及所有应用
+  chmod +x $WORK_DIR/{grpcwebproxy,cloudflared,nezha-agent,*.sh}
+
+fi
 
 # 运行 supervisor 进程守护
 supervisord -c /etc/supervisor/supervisord.conf
